+++ nested_admin/__init__.py
@@ -15,7 +15,7 @@ import django
 import django.forms.formsets
 import monkeybiz
 
-__version__ = '3.3.0'
+__version__ = '3.3.1'
 
 # import mapping to objects in other modules
 all_by_module = {